Add quarantine check and force reinstall for brew packages
The code has been updated to add a quarantine check when installing brew packages. If a file is quarantined, the existing package will be uninstalled and reinstalled with the --force flag. This provides a more robust approach to handle issues related to file quarantines while installing packages. The "makemkv" package is now being uninstalled before the installation process for other packages begins.